@@ -1,10 +1,8 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
-from setuptools import setup
 
-# To use a consistent encoding
-from codecs import open
 from os import path
+
+from setuptools import setup
 
 here = path.abspath(path.dirname(__file__))
 
@@ -13,11 +11,11 @@ with open(path.join(here, "ffmpeg_normalize", "_version.py")) as version_file:
     version = eval(version_file.read().split("=")[1].strip())
 
 # Get the long description from the README file
-with open(path.join(here, "README.md"), encoding="utf-8") as f:
+with open(path.join(here, "README.md")) as f:
     long_description = f.read()
 
 # Get the history from the CHANGELOG file
-with open(path.join(here, "CHANGELOG.md"), encoding="utf-8") as f:
+with open(path.join(here, "CHANGELOG.md")) as f:
     history = f.read()
 
 setup(
@@ -31,16 +29,20 @@ setup(
     url="https://github.com/slhck/ffmpeg-normalize",
     packages=["ffmpeg_normalize"],
     include_package_data=True,
+    package_data={
+        "ffmpeg_normalize": ["py.typed"],
+    },
     install_requires=[
-        "tqdm>=4.38.0",
-        "colorama>=0.4.3",
-        "ffmpeg-progress-yield>=0.1.2",
+        "tqdm",
+        "colorama",
+        "ffmpeg-progress-yield",
+        "colorlog",
     ],
     license="MIT",
     zip_safe=False,
     keywords="ffmpeg, normalize, audio",
     classifiers=[
-        "Development Status :: 4 - Beta",
+        "Development Status :: 5 - Production/Stable",
         "Intended Audience :: Developers",
         "Topic :: Multimedia :: Sound/Audio",
         "Topic :: Multimedia :: Sound/Audio :: Analysis",
@@ -48,12 +50,12 @@ setup(
         "License :: OSI Approved :: MIT License",
         "Natural Language :: English",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.6",
-        "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
     ],
-    python_requires=">=3.6",
+    python_requires=">=3.8",
     entry_points={
         "console_scripts": ["ffmpeg-normalize = ffmpeg_normalize.__main__:main"]
     },
