@@ -1,5 +1,18 @@
+from ._errors import FFmpegNormalizeError
 from ._ffmpeg_normalize import FFmpegNormalize
 from ._media_file import MediaFile
+from ._streams import AudioStream, MediaStream, SubtitleStream, VideoStream
 from ._version import __version__
 
-__all__ = ["FFmpegNormalize", "MediaFile", "__version__"]
+__module_name__ = "ffmpeg_normalize"
+
+__all__ = [
+    "FFmpegNormalize",
+    "FFmpegNormalizeError",
+    "MediaFile",
+    "AudioStream",
+    "VideoStream",
+    "SubtitleStream",
+    "MediaStream",
+    "__version__",
+]
