@@ -1,1 +1,1 @@
-__version__ = "1.22.5"
+__version__ = "1.26.6"
